build docker refactor & utils creation
